move IPlugin and ICompiler to src/client directory
