+++ src/client/plugin.ts
@@ -42,6 +42,33 @@ export interface ICompiler {
     renderMarkdown(markdown: string): string;
 }
 
+/**
+ * A Documentalist plugin is an object with a `compile(files, compiler)` function
+ * that returns a data object (or a promise for that object).
+ *
+ * A common pattern is to define a class that implements this interface and use the constructor
+ * to accept options.
+ *
+ * ```ts
+ * import { ICompiler, IFile, IPlugin } from "documentalist/client";
+ *
+ * export interface MyData {
+ *     pluginName: { ... }
+ * }
+ *
+ * export interface MyOptions {
+ *     ...
+ * }
+ *
+ * export class MyPlugin implements IPlugin<MyData> {
+ *     public constructor(options: MyOptions = {}) {}
+ *
+ *     public compile(files: IFile[], compiler: ICompiler) {
+ *         return files.map(transformToMyData);
+ *     }
+ * }
+ * ```
+ */
 export interface IPlugin<T> {
-    compile(files: IFile[], doc: ICompiler): T | Promise<T>;
+    compile(files: IFile[], compiler: ICompiler): T | Promise<T>;
 }
