+++ src/compiler.ts
@@ -7,8 +7,7 @@
 
 import * as yaml from "js-yaml";
 import * as marked from "marked";
-import { IBlock, IHeadingTag, StringOrTag } from "./client";
-import { ICompiler } from "./plugins";
+import { IBlock, ICompiler, IHeadingTag, StringOrTag } from "./client";
 
 /**
  * Matches the triple-dash metadata block on the first line of markdown file.
