+++ src/documentalist.ts
@@ -8,8 +8,8 @@
 import * as fs from "fs";
 import * as glob from "glob";
 import * as path from "path";
+import { IFile, IPlugin } from "./client";
 import { Compiler, ICompilerOptions } from "./compiler";
-import { IFile, IPlugin } from "./plugins";
 
 export interface IApi<T> {
     /**
