+++ src/plugins/typescript/index.ts
@@ -6,8 +6,7 @@
  */
 
 import { Application, ContainerReflection, ReflectionKind } from "typedoc";
-import { ITypescriptPluginData } from "../../client";
-import { ICompiler, IFile, IPlugin } from "../plugin";
+import { ICompiler, IFile, IPlugin, ITypescriptPluginData } from "../../client";
 import { isInternal, visitorExportedClass, visitorExportedInterface } from "./visitors";
 
 export interface ITypescriptPluginOptions {
