+++ src/plugins/typescript/visitors.ts
@@ -16,6 +16,7 @@ import {
 } from "typedoc";
 import { Comment } from "typedoc/dist/lib/models/comments/comment";
 import { DefaultValueContainer } from "typedoc/dist/lib/models/reflections/abstract";
+import { ICompiler } from "../../client/plugin";
 import {
     ITsClass,
     ITsInterface,
@@ -25,7 +26,6 @@ import {
     ITsProperty,
     Kind,
 } from "../../client/typescript";
-import { ICompiler } from "../plugin";
 import { resolveSignature, resolveTypeString } from "./typestring";
 
 export type Renderer = ICompiler["renderBlock"];
